langchain/knowledge-base-demo/requirements.txt
# FastAPI 后端依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
cors==1.0.1

# LangChain 相关
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# 数据处理
pydantic==1.10.12
python-dotenv==1.0.0

# 日志和工具
loguru==0.7.2
python-dateutil==2.8.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1